ipacm: Add back the ipanat library
Just to satisfy the dependency of the build without kernelplatform
and avoid missing ipanat library error. But libipanat from
techpack/dataipa is getting built and shipped with the prebuilt
kernel modules as it is and is getting used instead of this.

This reverts commit 4f926152425bd2d51e23d925af125778250a4357.
This repicks commit 856dafa524c5b7c3aa07ca2e83bd5509d10a5c87.

Change-Id: Ifcccb370d58e7301cff863908c1e10b24a6f7ed8